examples/tel/requirements.txt
python-telegram-bot==20.0a1
httpx
pydantic[dotenv]
